Only wrap the low-level methods once.
Rework the shimming so that there can be multiple agent run starts
and stops from a single process runtime while only adding the low-
level instrumentation once. My goal was to increase the testability
of the agent, but a byproduct is that this is a slightly cleaner
architectural approach overall.