Validate that Express instrumentation is working.
Writing the tests for Express was easy; getting them all to pass was
incredibly hard. In the end, there was a nasty and subtle bug related to
an unintentional closure being created around module._load that would
cause it to retain a reference to an outdated agent.
shimmer.preserveMethod tries its best to be idempotent, and it fixed
this problem, but first I needed to figure out where it needed to go.

There's still a lot of overhead around module loading, some of which is
maybe unnecessary. That's a big TODO.