Only need-to-know, strictly hush-hush.
There was a bunch of code that was overly generic for manipulating the
call stack. If it ever makes sense to split this off into a more general
library, it can be rewritten to be more general again, but for now,
keeping it simple makes for an easier to use API and less verbiage
everywhere.