Fix errors in agent event listener binding.
Because the agent directly manages its metrics now, several of the event
listeners need an additional level of indirection so they trigger
updates of the *current* normalization or renaming rules, rather than
those associated with agent's *initial* metrics. This would have been a
pain to debug, so I'm glad I caught it.