Eliminate the spectacularly bad idea of shimming multiple agents.
I thought it might be neat to be able to support multiple applications
within the same process, long ago, and generalized the transaction
propagation mechanism to support multiple agents being in play. This
incurred a bunch of overhead for no appreciable benefit and can be
easily enough re-added should we ever have the astoundingly poor fortune
to need it again.