Adding HTTP instrumentation no longer crashes Restify (closes: NA-47)
Before the code was assuming that http.createServer would always be
called with a callback for when the server is ready, which isn't a safe
assumption for a lot of use cases for the HTTP class.

However, now the tests for Express and Restify stomp on each other, so
two of the integration tests fail, and the test suite will need to be
rewritten with tap so that they're each run in independent subprocesses.