Added a set of tests exploring a new strategy for threading context.
Instead of using the costly and messy pervasive shim (and abusing the
hell out of the call stack and generating stacktraces all over the place),
use a tiny bit of global context as well as some very careful patterns
for putting state into and out of that context. Just read the file, as
it's pretty much self-contained.