Streamlined use of promises.
I'd forgotten that you can't just reuse a promise chain, as its final
resolved value is immutable. Also, should probably perform the final
data checks as the test user instead of as MySQL root. Split apart
the end and connect events and renamed everything, which will with luck
make the flow of the initialization fairly self-explanatory.