WIP: Versioned Express test refactor, added error-catcher tests.
This is the very rare occasion on which I commit failing tests, because
the scope of the changes is getting pretty huge, and I wanted to get
this refactor in before I started pulling Transaction and other tracer
classes in to get the tests passing.

A byproduct of pushing the Express and Connect tests into the versioned
tests is that the core modulee no longer has dev dependencies on Express
or Connect -- the examples will need updates to their package.json
files, but that can come later.

Also includes some documentation cleanup in shimmer and a new, dangerous
helper method.