Fix Connect handler wrapping and make error interception heuristic.
This is probably best considered a work in progress and needs better
tests. Copying the longish comment from wrapUse for context:

Give the error tracer a better chance of intercepting errors by putting
it before the first error handler (a middleware that takes 4 parameters,
in Connect's world). Error handlers tend to be placed towards the end of
the middleware chain and sometimes don't pass errors along. Don't just
put the interceptor at the beginning because we want to allow as many
middleware functions to execute as possible before the interceptor is
run, to increase error coverage.

NOTE: This is heuristic, and works because interceptor propagates errors
      instead of terminating the middleware chain. Ignores routes.