For metric names, treat '?' and ';' equivalentlty.
The most common use cases for semicolons in URLs, aside from an
alternative delimiter for query parameters (as per W3C recommendation,
to avoid having to escape '&'), is to act as a pseudo-query parameter
for session tracking. This can cause metric grouping explosions, so
take the normalized URL and strip anything after ';', if found.