Replace parameter 'agent_run_id' with 'run_id'.
This whole time the Node agent has been using the wrong parameter to
pass the run ID to the collector when sending data (the docs are
inconsistent, saying 'run_id' in one place but 'agent_run_id'
elsewhere -- they're next). For whatever reason, this has been working
fine, but now is causing the collector shark to break, so now is the
time to fix it.