error tracer: made lame hack less lame.
We want to kill the transaction on errors, so ending it makes sense. But
do we also want to mark the transaction as invalid somehow so that the
trace doesn't get sent back? Or do we want to wrap the transaction trace
up in the error trace to provide context? Decision pending.