metrics: merge on failed transmission.
There's been a regression lurking for a while now that only got
exercised when the agent switched to Object.keys from an old-style
for...in enumeration of the properties on the object. This exposed
that the object being passed to Agent.prototype.mergeMetrics was not, in
fact, a metrics object, but the pre-serialization JSON array expected by
the collector.

Also added nock because Sinon doesn't have a way to easily mock HTTP
requests from Node. Nock's nice! Thanks, Pedro!