tracer: web segment name must use normalized URI.
For display to work properly in RPM, the name of the top-level trace
segment must match the name of the WebTransaction metric name. Before,
metrics were being recorded with the normalized URI name, but the
segment was getting created before normalization could occur. The
current solution is kind of a hack, but that's fine, because all this
will need to be rationalized as part of fixing up the outstanding metric
grouping issues anyway.