instrumentation: set Redis key as a parameter.
This is a pretty quick and dirty hack; the test coverage is minimal and
it doesn't deal with any multi-key operations at all right now. But it
does work out of the box for the common cases, and the tests that do
exist are using the live client and a real server.

Modified the server to not dump an RDB file, and the default config
already doesn't create AOF logs.