metrics: record exclusive times at each level
This is a very fragile change; "exclusive time" is a much fuzzier
concept in Node, where it's hard to tell the difference between when a
computation is suspended and when it's consuming lots of CPU, and also
when many asynchronous processes can be in flight simultaneously.

This calculation happened only at the trace level before, which is
clearly not where it belongs. The version in use right now isn't
recursive, which means that different levels of execution can report
inconsistent exclusive times. This may need to be tuned for the UI, but
it's actually a pretty close reflection of reality.

Perhaps more problematic is the fact that metrics recordings may be
happening before the transaction is complete, which means that the
recorders may be operating on inconsistent data. This is something that
may need to be reworked as part of MGI, so leaving it alone for now.