errors: clean up monitoring for Node 0.8
Tracing errors in 0.6 requires the use of
process.on('uncaughtException'). Tracing errors in 0.8 can be done with
domains, which are built on top of process.on('uncaughtException').
Domains in 0.10 are implemented closer to Node core. Writing tests
against error handling that are agnostic to the Node version is
therefore pretty hard, and a little hacky -- witness the latches in the
tests in this commit.

The good news is that error handling is simpler overall in the agent
than it was before, but another pass may be necessary to capture
everything.