metrics: record metrics on the correct scope.
The previous system, in addition to being overly complicated, passed the
segment's name as the scope, which was wrong the vast majority of the
time. To allow individual trace segments to close before the scope of
the transaction is named (to allow for the URI / transaction name to
be set late in the transaction-processing lifecycle), processing of the
segments has been shifted to the tracer itself. This means that no
segment metrics will be recorded until the trace is ended.