metrics: RenameRules -> MetricMapper
The idea of "renaming" a metric to an ID is very confusing, especially
when there's metric normalization around to further muddy the waters.
Since what's actually happening is mapping a name or name and scope pair
to a short ID, call it a metric mapper.

Also, don't know what's up with all this indirection. Metrics
collections need a reference to a mapper so they can exist independently
of the agent or transaction, but we should always be accessing the
mapper through the agent when possible, not another object accessible
via the agent.