api: added two naming functions
Here are the first two methods on the exposed API (not yet wired up to
the agent, because there are some probably unrelated test failures to
clean up first). Doing this provided an opportunity to do some
simplification and refactoring of the naming and normalization functions
related to the API, along with putting in place the backstop.