transaction: scopes are only for metrics
I was so confused about this for so long, and here's where I ended up:
transactions have names, which can be set from the URL, by the naming
API, or by naming rules. There is a class of metrics that are attached
to a specific transaction via a scope, and that scope is identified by
the transaction's name. I hope you find this clarification useful.