errors: use actual transaction naming
The error tracer was doing its own normalization, which was a bad idea
for a whole bunch of obvious reasons. Replaced with the actual
transaction naming call, with some safeguards added to ensure that
if the traced error is recovered, the transaction will still end up with
the correct name.