normalizer: minor deoptimization fix
Needed to add some tests to ensure I wasn't screwing things up with this
"minor" change, though. As well as start the ball rolling on having the
tests enumerate what properties are in use by the objects in the system,
for understanding what's being done to prevent deoptimization, if
nothing else.