config: config can now use the logger
Had to do a bit of a two-step to resolve the cyclical dependency, so now
the configuration has a method that tells it to call a method on the
logger that returns the current logger, and the logger bootstraps itself
to allow the config to log before it's finished bootstrapping.
Craziness, but it works.