config: SSC support for metric renaming rules
Also changed Metrics to require a mapper and a normalizer at creation.
The old way of defaulting everything made for tidy tests, but didn't
really model how the metrics were being used in the agent. Adding the
normalizer means that metrics collections also carry around the config,
but that also means they're using the live values from the config,
except for apdex_t, which needs to be consistent for a given metrics
collection over a given harvest cycle.