config: env overrides file overrides defaults
Instead of trying to be clever and concatenate list values together,
which made it impossible to override default values in lists (e.g. the
HTTP status codes to ignore), just overwrite them when overriding. So
the configuration file values will overwrite the defaults, and the
environment variables will in turn overwrite the config file and
defaults. The only exception is configuration coming from the collector,
which is additive. This is actually a little simpler, but there's still
a lot (too much?) going on around these multiple levels of
configuration.