connection: retry logic works again
Sometime back in January, I think, I broke the logic that caused the
agent to repeatedly try to connect to Monis Agent (with backoffs between
attempts). This stuff is pretty tricky to test in situ, which is why the
new test case involves time travel (a feature I didn't know Sinon
supported).