errors: remove dominion
Instead modify the CLS async listener to include an error tracer. Unit
tests are all passing, and the two integration tests that are failing
highlight behavior in CLS that requires some analysis to properly
address.

The new strategy creates significantly fewer objects and is measurably
faster, just running the tests. If we can figure out how to make it
work, we will have a major win.