urltils: put parameter copying in one place
awinterman pointed out on twitter that relying on a Node core private
API call (util._extend()) was probably unsafe and a bad example for the
children. Not only that, it was duplicated all over the place, not
tested, and allocated more objects than necessary. Now there's one
function that will handle this everywhere, dedicated unit tests, and a
simpler overall implementation.