config: simplify logger-config dance
Dealing with cyclic dependencies is hard, but having a bunch of largely
redundant boilerplate in many places is gross. This still doesn't solve
the basic problem of config and logger depending directly on each other,
but without bunyan supporting the changing of log locations and types on
the fly, this is about as clean as it can be.

Also cleans out unnecessary layer of config wrapping for initialize. If
you're passing a configuration object in, it really shouldn't be
necessary to have that wrapped up inside the object in a property named
'config'.