collector: don't treat license exception as fatal
Sometimes the collector gets confused temporarily about whether a
license key is or isn't valid; to prevent this from stopping data
reporting from a running agent until the application is restarted, treat
LicenseKeyException as just another transient error.

To keep the tests working, there needs to be some way of observing the
state of the agent, and signalling when it's started stopped. I'd really
like to get rid of this mechanism eventually (because it's brittle), but
it does add a useful layer of debuggability to the module and connection
progress for now.