trace: make aggregator harvest less racy
There's a window between which the transaction trace harvest begins and
when confirmation comes back that the trace has been successfully
assimilated. If a slower trace has come along in this window and
replaced the existing one, and the delivery was successful, don't blow
it away during the harvest cycle.