remote-method: Fix proxy support when using proxy_host and proxy_port
When using proxy_host and proxy_port, rather than proxy, the wrong
location was used to connect to the collector. This bit of logic was
left over after we updated proxy support in version 1.8.0.

The change in proxy-api-connection.tap.js also ensures we have a test
for both proxy and proxy_host/proxy_port.