CollectorAPI: pass in the full agent instead of just the metrics
Although we actually only need the metrics, there's a problem in how
we define an agent's `CollectorAPI` and `MetricAggregator`. The
agent's collector API is defined first and then the metrics require
this CollectorAPI object. This means that at the moment of definition
of the collector API, the metrics aren't defined yet.

But since we don't actually invoke the metrics until later, it's okay
to pass in the whole agent into the collector API and just call the
metrics from the agent object, later, after they've been defined.