remote-method: make deflate the non-default option
If something goes wrong with the config (say, the user typoes
"deflate", or the config is just somehow missing altogether), we need
to default to gzip compression.

One of the tests is running without any config at all and was still
defaulting to deflate compression sizes, demonstrating the need for
this change.