NEWRELIC-6211: (winston) also instrument winston.loggers.add
There is a different part of the Winston API that our current
instrumentation simply misses. It's possible to create a logger by
directly calling `createLogger`, but it's also possible to add new
loggers to the list of loggers by doing `winston.loggers.add` with
almost the same syntax.

This factors out a bit of the logic so that both methods of creating
winston loggers are instrumented.